Address FFI review: CFNumberType width + unified-memory copy gating
Two confirmed findings from the adversarial FFI review:

- CFNumberGetValue's `theType` is CFIndex (c_long, 64-bit on macOS), not
  c_int. Passing a 32-bit value through a 64-bit ABI slot is UB at the FFI
  boundary (benign only by codegen accident). Declare it c_long and make
  SINT64 a c_long constant.
- The AI view's "unified memory" line keyed only on mem_total == 0, so a
  Linux/sysfs GPU that failed to report VRAM total would be mislabeled as
  unified memory. Gate the wording on the Apple GPU name; other zero-total
  GPUs now show "--".

Re-verified live on Apple Silicon (utilization still reads).

+++ src/metrics/gpu.rs
@@ -310,7 +310,7 @@ fn query_nvidia_procs() -> Vec<GpuProc> {
 #[cfg(target_os = "macos")]
 mod apple {
     use std::ffi::{c_void, CString};
-    use std::os::raw::{c_char, c_int};
+    use std::os::raw::{c_char, c_int, c_long};
 
     type CFTypeRef = *const c_void;
     type CFStringRef = *const c_void;
@@ -319,14 +319,16 @@ mod apple {
     type IoObject = u32;
 
     const UTF8: u32 = 0x0800_0100; // kCFStringEncodingUTF8
-    const SINT64: c_int = 4; // kCFNumberSInt64Type
+                                   // kCFNumberSInt64Type. CFNumberType is backed by CFIndex (c_long, 64-bit on
+                                   // macOS), so this must be c_long — a c_int here is an FFI ABI mismatch.
+    const SINT64: c_long = 4;
     const NULL_ALLOC: CFAllocatorRef = std::ptr::null();
 
     #[link(name = "CoreFoundation", kind = "framework")]
     extern "C" {
         fn CFStringCreateWithCString(a: CFAllocatorRef, s: *const c_char, enc: u32) -> CFStringRef;
         fn CFDictionaryGetValue(d: CFDictionaryRef, k: *const c_void) -> *const c_void;
-        fn CFNumberGetValue(n: *const c_void, t: c_int, v: *mut c_void) -> bool;
+        fn CFNumberGetValue(n: *const c_void, t: c_long, v: *mut c_void) -> bool;
         fn CFRelease(cf: CFTypeRef);
     }
